papers/iros2022/requirements.txt
# this should be run from the main enpheeph project directory, so that -e . installs
# the enpheeph package, e.g. pip install -r papers/iros2022/requirements.txt
# tested with Python 3.9.7
# it is suggested to use conda and install the following dependencies from conda-forge
# the main problem is with the cudatoolkit as well as cupy, which is difficult
# to build properly using pip
# use cuda 10.1 if cuda-dev is required later on, as the cuda10.2 are included in
# pytorch
# cudatoolkit=10.2
# cupy=9.6.0
# the compilers might not be needed, it depends on which packages need to be built
# gcc_linux-64
# gxx_linux-64
# if unsure just install them
# NOTE: cudatoolkit-dev=10.2 DOES NOT EXIST


-e .[full]
# to upgrade packages only if needed to match the dependencies
# must be used as command-line options
# --upgrade --upgrade-strategy only-if-needed

# for PyTorch 1.9.1
-f https://download.pytorch.org/whl/torch/

lightning-flash[all]==0.5.2
norse==0.0.7.post1
pytorch-lightning[extra]==1.5.2
tonic==1.0.13
torch==1.9.1+cu102
torchaudio==0.9.1
# torchinfo is required to run the random injections
torchinfo==1.6.3
torchvision==0.10.1
